requirements.txt
pgpy13
psutil
prompt_toolkit
click
cryptography
python-gnupg
keyring